+++ src/commands/command.ts
@@ -26,6 +26,7 @@ export class Commands {
 		this.registerReloadConfigCmd(context, accountViewProvider, contractViewProvider);
 		this.syncBeakerTomlCmd(context);
 		this.createNewCWNotebookCmd(context);
+		this.encodeToBase64Cmd(context);
 		//this.registerRecordCWCmd(context);		
 	}
 
@@ -147,6 +148,23 @@ export class Commands {
 				cells.push(new vscode.NotebookCellData(vscode.NotebookCellKind.Code, '{\r\n    \"increment\": {}\r\n}', Constants.LANGUAGE_JSON));
 				return cells;
 			}
+		});
+		context.subscriptions.push(disposable);
+	}
+
+	private static encodeToBase64Cmd(context: vscode.ExtensionContext) {
+		let disposable = vscode.commands.registerCommand('cosmy-wasmy.encodeToBase64', (item: vscode.Uri) => {
+			let activeTextEditor = vscode.window.activeTextEditor;
+			if (activeTextEditor.document.uri.path != item.path) {
+				vscode.window.showErrorMessage(vscode.l10n.t("Something went wrong in fetching the selected text. Please try again!"));
+				return;
+			}
+			const selection = activeTextEditor.selection;
+			activeTextEditor.edit(function(e) {
+				const selectedText = activeTextEditor.document.getText(new vscode.Range(selection.start, selection.end));
+				let b = Buffer.from(selectedText);
+				e.replace(selection, b.toString("base64"));
+			})
 		});
 		context.subscriptions.push(disposable);
 	}